move monorepo changelog to npm plugin
