+++ src/github-release.ts
@@ -9,6 +9,7 @@ import generateReleaseNotes, {
   IExtendedCommit,
   normalizeCommits
 } from './log-parse';
+import { IAutoHooks } from './main';
 import SEMVER, { calculateSemVerBump } from './semver';
 import execPromise from './utils/exec-promise';
 import { dummyLog, ILogger } from './utils/logger';
@@ -80,6 +81,7 @@ const writeFile = promisify(fs.writeFile);
 export default class GitHubRelease {
   public readonly releaseOptions: IGitHubReleaseOptions;
 
+  private readonly hooks: IAutoHooks;
   private readonly logger: ILogger;
   private readonly github: GitHub;
   private readonly changelogTitles: { [label: string]: string };
@@ -91,8 +93,10 @@ export default class GitHubRelease {
     releaseOptions: IGitHubReleaseOptions = {
       skipReleaseLabels: []
     },
-    logger: ILogger = dummyLog()
+    logger: ILogger = dummyLog(),
+    hooks: IAutoHooks
   ) {
+    this.hooks = hooks;
     this.versionLabels = releaseOptions.versionLabels || defaultLabels;
     this.logger = logger;
     this.releaseOptions = releaseOptions;
@@ -131,18 +135,13 @@ export default class GitHubRelease {
     const commits = await this.getCommits(from, to);
     const project = await this.github.getProject();
 
-    await Promise.all(
-      commits.map(async commit => {
-        commit.packages = await this.github.changedPackages(commit.hash);
-      })
-    );
-
     return generateReleaseNotes(commits, this.logger, {
       owner: this.github.options.owner,
       repo: this.github.options.repo,
       baseUrl: project.html_url,
       jira: this.releaseOptions.jira,
       versionLabels: this.versionLabels,
+      modifyChangelog: this.hooks.modifyChangelog,
       changelogTitles: {
         ...defaultChangelogTitles,
         ...this.changelogTitles
